+++ src/main/java/com/hsbc/roboadvisor/model/Fund/TransactionDetail.java
@@ -1,12 +1,12 @@
 package com.hsbc.roboadvisor.model.Fund;
 
-import com.hsbc.roboadvisor.model.Common.CurrencyAmount;
-import com.hsbc.roboadvisor.model.Recommendation.TransactionType;
-
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+
+import com.hsbc.roboadvisor.model.Common.CurrencyAmount;
+import com.hsbc.roboadvisor.model.Recommendation.TransactionType;
 
 public class TransactionDetail {
 
@@ -15,7 +15,7 @@ public class TransactionDetail {
     private TransactionType action;
 
     @NotNull
-    private Integer fundId;
+    private Long fundId;
 
     @NotNull
     @Min(value = 0)
@@ -32,11 +32,11 @@ public class TransactionDetail {
         this.action = action;
     }
 
-    public Integer getFundId() {
+    public Long getFundId() {
         return fundId;
     }
 
-    public void setFundId(Integer fundId) {
+    public void setFundId(Long fundId) {
         this.fundId = fundId;
     }
