+++ src/main/java/com/hsbc/roboadvisor/repository/PortfolioRepository.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Repository;
 import com.hsbc.roboadvisor.model.PortfolioPreference.PortfolioPreference;
 
 @Repository
-public interface PortfolioRepository extends CrudRepository<PortfolioPreference, Integer>
+public interface PortfolioRepository extends CrudRepository<PortfolioPreference, Long>
 {
 
-    PortfolioPreference findByPortfolioId(Integer portfolioId);
+    PortfolioPreference findByPortfolioId(Long portfolioId);
 
-    Boolean existsByPortfolioId(Integer portfolioId);
+    Boolean existsByPortfolioId(Long portfolioId);
     
 }
