+++ src/main/java/com/hsbc/roboadvisor/service/FundRecommendationService.java
@@ -104,8 +104,8 @@ public class FundRecommendationService
             throw new BadRequestException("No funds exists for the requested category");
         }
 
-        Map<Integer, Double> fundScoreMap = new HashMap<>();
-        Map<Integer, Transaction> fundTransactionMap = new HashMap<>();
+        Map<Long, Double> fundScoreMap = new HashMap<>();
+        Map<Long, Transaction> fundTransactionMap = new HashMap<>();
 
         for (Fund fund : fundsOfCategory) {
             Double score = calculatePerformanceScore(fund);
@@ -117,13 +117,13 @@ public class FundRecommendationService
 
         _logger.info("Scores per fund {}", fundScoreMap.toString());
 
-        Map<Integer, Double> sortedFundScoreMap = fundScoreMap.entrySet().stream()
-                .sorted(Map.Entry.<Integer, Double>comparingByValue().reversed())
+        Map<Long, Double> sortedFundScoreMap = fundScoreMap.entrySet().stream()
+                .sorted(Map.Entry.<Long, Double>comparingByValue().reversed())
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (e1, e2) -> e1, LinkedHashMap::new));
 
         List<Transaction> transactionList = new ArrayList<>();
 
-        for (Integer fundId : sortedFundScoreMap.keySet()) {
+        for (Long fundId : sortedFundScoreMap.keySet()) {
             transactionList.add(fundTransactionMap.get(fundId));
         }
